fix(dip): add data-in-models
