+++ .github/workflows/pull-request.yml
@@ -63,6 +63,7 @@ jobs:
             data-in-pipeline:
               - 'data-in-pipeline/**'
               - 'api/**'
+              - 'data-in-models/**'
               - 'uv.lock'
               - pyproject.toml
