+++ data-in-models/README.md
@@ -1,0 +1,42 @@
+# Data-in models
+
+## Installation
+
+### Within this repo
+
+```bash
+# from the root of the repo
+uv add --project {service} data-in-models
+```
+
+- add the dependency to the `changes` step in `merge-to-main.yml` and
+  `pull-request.yml` workflows
+
+  ```yml
+  # ...
+  changes:
+    # ...
+    steps:
+      #...
+      id: filter
+      with:
+        filters: |
+          families-api:
+            - 'families-api/**'
+            - 'api/**'
+            - 'data-in-models/**' # <= this line
+            - 'uv.lock'
+            - pyproject.toml
+    # ...
+  ```
+
+- `COPY` the folder to your `Dockerfile`
+
+  ```dockerfile
+  COPY ./data-in-models /app/data-in-models/
+  ```
+
+### From another repo
+
+> [!NOTE]  
+> 🚧 We are currently working on this 🚧
