+++ data-in-pipeline/app/navigator_document_etl_pipeline.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 
+from data_in_models.models import Document
 from prefect import flow, task
 from prefect.runtime import flow_run, task_run
 from returns.pipeline import is_successful
@@ -11,7 +12,7 @@ from app.extract.connectors import NavigatorConnector, NavigatorDocument
 from app.extract.enums import CheckPointStorageType
 from app.identify.navigator_document import identify_navigator_document
 from app.load.aws_bucket import upload_to_s3
-from app.models import Document, ExtractedEnvelope, Identified
+from app.models import ExtractedEnvelope, Identified
 from app.pipeline_metrics import Operation, PipelineType, Status
 from app.transform.navigator_document import transform_navigator_document
