+++ data-in-pipeline/app/navigator_family_etl_pipeline.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from typing import Literal
 
+from data_in_models.models import Document
 from prefect import flow, task
 from prefect.runtime import flow_run, task_run
 from returns.result import Failure, Success
@@ -16,7 +17,7 @@ from app.extract.enums import CheckPointStorageType
 from app.identify.navigator_family import identify_navigator_family
 from app.load.aws_bucket import upload_to_s3
 from app.load.load import load_to_db
-from app.models import Document, ExtractedEnvelope, Identified
+from app.models import ExtractedEnvelope, Identified
 from app.pipeline_metrics import ErrorType, Operation, PipelineType, Status
 from app.run_migrations.run_migrations import run_migrations
 from app.transform.navigator_family import (
