+++ data-in-pipeline/app/transform/navigator_family.py
@@ -1,3 +1,11 @@
+from data_in_models.models import (
+    Document,
+    DocumentDocumentRelationship,
+    DocumentLabelRelationship,
+    DocumentWithoutRelationships,
+    Item,
+    Label,
+)
 from returns.result import Failure, Result, Success
 
 from app.bootstrap_telemetry import get_logger, log_context
@@ -7,15 +15,7 @@ from app.extract.connectors import (
     NavigatorFamily,
 )
 from app.geographies import geographies_lookup
-from app.models import (
-    Document,
-    DocumentDocumentRelationship,
-    DocumentLabelRelationship,
-    DocumentWithoutRelationships,
-    Identified,
-    Item,
-    Label,
-)
+from app.models import Identified
 from app.transform.models import CouldNotTransform, NoMatchingTransformations
 
 mcf_projects_corpus_types = [
