+++ data-in-pipeline/pyproject.toml
@@ -5,6 +5,7 @@ readme = "README.md"
 requires-python = ">=3.12,<3.14"
 dependencies = [
   "api",
+  "data-in-models",
   "components>=1.0.3",
   "prefect[aws]>=3.4",
   "pulumi>=3.202.0",
@@ -29,6 +30,7 @@ dependencies = [
 [tool.uv.sources]
 api = { workspace = true }
 components = { git = "https://github.com/climatepolicyradar/infra", subdirectory = "components", tag = "v1.0.5" }
+data-in-models = { workspace = true }
 
 
 [dependency-groups]
