+++ pyproject.toml
@@ -8,6 +8,7 @@ dependencies = []
 [tool.uv.workspace]
 members = [
   "api",
+  "data-in-models",
   "concepts-api",
   "families-api",
   "geographies-api",
@@ -25,6 +26,7 @@ extraPaths = ["api/src"]
 [tool.pyright.executionEnvironments]
 root = [
   "api/src",
+  "data-in-models",
   "concepts-api",
   "families-api",
   "geographies-api",
